add: add account admin and update inventory service

+++ src/Aseta.API/Controllers/InventoryController.cs
@@ -1,6 +1,7 @@
+using System.Security.Claims;
+using Aseta.Application.Abstractions.Checkers;
 using Aseta.Application.Abstractions.Services;
 using Aseta.Application.DTO.Inventory;
-using Aseta.Domain.Entities.Items;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -8,48 +9,83 @@ namespace Aseta.API.Controllers;
 
 [Route("[controller]")]
 [ApiController]
-public class InventoryController(IInventoryService inventoryService, IAuthorizationService authorizationService) : ControllerBase
+public class InventoryController(IInventoryService inventoryService,
+    ICheckingAccessPolicy checkingAccessPolicy,
+    ICheckingLockoutUser checkingLockoutUser) : ControllerBase
 {
     private readonly IInventoryService _inventoryService = inventoryService;
-    private readonly IAuthorizationService _authorizationService = authorizationService;
+    private readonly ICheckingAccessPolicy _checkingAccessPolicy = checkingAccessPolicy;
+    private readonly ICheckingLockoutUser _checkingLockoutUser = checkingLockoutUser;
 
     [HttpPost("create-item")]
     public async Task<ActionResult> CreateItem(CrateItemRequest request)
     {
-        var authorizationResult = await _authorizationService.AuthorizeAsync(
-            User,
-            request.InventoryId,
-            "CanEditInventory"
-        );
+        if (!await _checkingAccessPolicy.CheckAsync(User, request.InventoryId, "CanEditInventory"))
+            return Forbid("User not has permission");
 
-        if (!authorizationResult.Succeeded)
-        {
-            return Forbid();
-        }
+        if (await _checkingLockoutUser.CheckAsync(User))
+            return Forbid("User is lockout");
 
-        await _inventoryService.AddItemAsync(request);
+        var userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value
+            ?? throw new Exception("User not found");
+
+        await _inventoryService.AddItemAsync(request, Guid.Parse(userId));
+        return Ok();
+    }
+
+    [HttpPut("update-item")]
+    public async Task<ActionResult> UpdateItem(UpdateItemRequest request)
+    {
+        if (!await _checkingAccessPolicy.CheckAsync(User, request.InventoryId, "CanEditInventory"))
+            return Forbid("User not has permission");
+
+        if (await _checkingLockoutUser.CheckAsync(User))
+            return Forbid("User is lockout");
+
+        var userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value
+            ?? throw new Exception("User not found");
+
+        await _inventoryService.UpdateItemAsync(request, Guid.Parse(userId));
         return Ok();
     }
 
     [HttpDelete("remove-item")]
     public async Task<ActionResult> RemoveItem(RemoveItemRequest request)
     {
+        if (!await _checkingAccessPolicy.CheckAsync(User, request.InventoryId, "CanEditInventory"))
+            return Forbid("User not has permission");
+
+        if (await _checkingLockoutUser.CheckAsync(User))
+            return Forbid("User is lockout");
+
         await _inventoryService.RemoveItemAsync(request);
         return Ok();
     }
 
-    [HttpDelete("remove-inventory")]
-    public async Task<ActionResult> RemoveInventory(RemoveInventoryRequest request)
+    [HttpDelete("remove-inventory{InventoryId}")]
+    public async Task<ActionResult> RemoveInventory(Guid InventoryId)
     {
-        await _inventoryService.RemoveInventoryAsync(request);
+        if (!await _checkingAccessPolicy.CheckAsync(User, InventoryId, "CanManageInventory"))
+            return Forbid("User not has permission");
+
+        if (await _checkingLockoutUser.CheckAsync(User))
+            return Forbid("User is lockout");
+
+        await _inventoryService.RemoveInventoryAsync(InventoryId);
         return Ok();
     }
 
+    [Authorize]
     [HttpPost("create-inventory")]
     public async Task<ActionResult> CreateInventory(CreateInventoryRequest request)
     {
-        await _inventoryService.CreateInventoryAsync(request);
+        if (await _checkingLockoutUser.CheckAsync(User))
+            return Forbid("User is lockout");
+
+        var userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value
+            ?? throw new Exception("User not found");
+
+        await _inventoryService.CreateInventoryAsync(request, Guid.Parse(userId));
         return Ok();
     }
 }
-
